requirements.txt
contourpy==1.0.7
coverage==7.0.5
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
plum-py==0.8.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
